feat: added the dockerfile, blueprints and updeated the services, app and models_db
