+++ docker-compose.yml
@@ -1,22 +1,39 @@
-version: "3.7"
+version: '3.8'
+
 services:
-  db:
-    image: postgres:latest
+  # PostgreSQL Database
+  postgres:
+    image: postgres:15
+    environment:
+      POSTGRES_DB: fit-db
+      POSTGRES_USER: postgres
+      POSTGRES_PASSWORD: docker
     ports:
       - "5432:5432"
-    command: postgres -c max_connections=200
-    user: postgres
-    environment:
-      - POSTGRES_DB=fit-db
-      - POSTGRES_USER=postgres
-      - POSTGRES_PASSWORD=docker
+    volumes:
+      - postgres_data:/var/lib/postgresql/data
+      - ./src/fit/db_init_scripts:/docker-entrypoint-initdb.d
     healthcheck:
-      test: ["CMD-SHELL", "pg_isready", "-d", "fit-db"]
-      interval: 5s
-      timeout: 30s
+      test: ["CMD-SHELL", "pg_isready -U postgres"]
+      interval: 10s
+      timeout: 5s
       retries: 5
-    networks:
-      - fit
 
-networks:
-  fit:
+  # Flask Application
+  app:
+    build: .
+    ports:
+      - "5000:5000"
+    environment:
+      - FLASK_ENV=development
+      - SQLALCHEMY_DATABASE_URL=postgresql://postgres:docker@postgres:5432/fit-db
+    depends_on:
+      postgres:
+        condition: service_healthy
+    volumes:
+      - ./src:/app/src
+      - ./tests:/app/tests
+    restart: unless-stopped
+
+volumes:
+  postgres_data:
