+++ main.py
@@ -3,3 +3,4 @@ from src.fit.app import run_app
 
 if __name__ == "__main__":
     run_app() 
+
