+++ src/fit/services/fitness_coach_service.py
@@ -1,6 +1,7 @@
 from typing import List, Tuple
-from ..models_db import ExerciseModel, MuscleGroupModel, exercise_muscle_groups
+from ..models_db import ExerciseModel, MuscleGroupModel, exercise_muscle_groups, UserExerciseHistoryModel
 from ..database import db_session
+from datetime import datetime, timedelta
 import random
 from time import time
 
@@ -25,9 +26,48 @@ def calculate_intensity(difficulty: int) -> float:
     # Convert difficulty (1-5) to intensity (0.0-1.0)
     return (difficulty - 1) / 4.0
 
-def request_wod() -> List[Tuple[ExerciseModel, List[Tuple[MuscleGroupModel, bool]]]]:
+def get_recent_exercises(user_email: str, days_back: int = 3) -> List[int]:
     """
-    Request a workout of the day (WOD).
+    Get exercise IDs that the user has performed in the last N days.
+    """
+    db = db_session()
+    try:
+        cutoff_date = datetime.utcnow() - timedelta(days=days_back)
+        
+        recent_exercises = db.query(UserExerciseHistoryModel.exercise_id).filter(
+            UserExerciseHistoryModel.user_email == user_email,
+            UserExerciseHistoryModel.performed_at >= cutoff_date
+        ).distinct().all()
+        
+        return [ex[0] for ex in recent_exercises]
+    finally:
+        db.close()
+
+def save_exercise_history(user_email: str, exercise_id: int, weight: float, reps: int):
+    """
+    Save the user's exercise to history.
+    """
+    db = db_session()
+    try:
+        history_entry = UserExerciseHistoryModel(
+            user_email=user_email,
+            exercise_id=exercise_id,
+            suggested_weight=weight,
+            suggested_reps=reps
+        )
+        
+        db.add(history_entry)
+        db.commit()
+    except Exception as e:
+        db.rollback()
+        raise e
+    finally:
+        db.close()
+
+def request_wod(user_email: str) -> List[Tuple[ExerciseModel, List[Tuple[MuscleGroupModel, bool]]]]:
+    """
+    Request a workout of the day (WOD) for a specific user.
+    Avoids exercises the user has done recently.
     Returns a list of tuples containing:
     - The exercise
     - A list of tuples containing:
@@ -39,11 +79,22 @@ def request_wod() -> List[Tuple[ExerciseModel, List[Tuple[MuscleGroupModel, bool
     
     db = db_session()
     try:
-        # Get all exercises with their muscle groups
-        exercises = db.query(ExerciseModel).all()
+        # Get exercises the user has done recently
+        recent_exercise_ids = get_recent_exercises(user_email)
+        
+        # Get all exercises, excluding recent ones
+        query = db.query(ExerciseModel)
+        if recent_exercise_ids:
+            query = query.filter(~ExerciseModel.id.in_(recent_exercise_ids))
+        
+        available_exercises = query.all()
+        
+        # If we don't have enough available exercises, fall back to all exercises
+        if len(available_exercises) < 6:
+            available_exercises = db.query(ExerciseModel).all()
         
         # Select 6 random exercises
-        selected_exercises = random.sample(exercises, 6) if len(exercises) >= 6 else exercises
+        selected_exercises = random.sample(available_exercises, 6) if len(available_exercises) >= 6 else available_exercises
         
         # For each exercise, get its muscle groups and whether they are primary
         result = []
@@ -62,6 +113,11 @@ def request_wod() -> List[Tuple[ExerciseModel, List[Tuple[MuscleGroupModel, bool
             muscle_groups = [(mg, is_primary) for mg, is_primary in stmt.all()]
             result.append((exercise, muscle_groups))
             
+            # Save to history (with random weight and reps for tracking)
+            weight = random.uniform(5.0, 50.0)
+            reps = random.randint(8, 15)
+            save_exercise_history(user_email, exercise.id, weight, reps)
+            
         return result
     finally:
-        db.close() 
+        db.close()
